requirements.txt
# requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
plotly>=5.20.0
anthropic>=0.25.0
openai>=1.30.0
google-generativeai>=0.5.0